refactor: extract common tab component to OptionsTabs
